pkg: sxmo: sxmo-utils: add PPP deviceprofile
The name of the speaker (used by amixer) seems to be different on the
PPP than the PinePhone. This includes the deviceprofile for the PPP
(which isn't in a release yet) and uses it to set the speaker.

This is a temporary fix until sxmo switches to pipewire / pulseaudio,
which will probably be in 1.9.

Fixes #340